Refactor `Program`, add method indirection.
The `Program` class was basically un-testable due to the static nature and number of cross connected functions. The first step in unwinding this mess is to add a layer of delegate indirection, which tests could then intercept as necissary.
